Responding to requested changes.
