+++ src/DataLib/fendl32B_retrofit/process_fendl3.2.py
@@ -1,9 +1,9 @@
 # Import packages
+import shutil
 import reaction_data as rxd
 import tendl_processing as tp
 import groupr_tools
-import file_handling as fh
-from pandas import concat
+from pandas import DataFrame
 
 def main():
     """
@@ -14,14 +14,14 @@ def main():
     TAPE21 = 'tape21'
 
     mt_dict = rxd.process_mt_data(rxd.load_mt_table('mt_table.csv'))
-    cumulative_data = tp.cumulative_data
 
-    for isotope, file_properties in fh.search_for_files().items():
+    cumulative_data = []
+    for isotope, file_properties in tp.search_for_files().items():
         element = file_properties['Element']
         A = file_properties['Mass Number']
         endf_path, pendf_path = file_properties['File Paths']
-        endf_path.rename(TAPE20)
-        pendf_path.rename(TAPE21)
+        shutil.copy(endf_path, TAPE20)
+        shutil.copy(pendf_path, TAPE21)
 
         material_id, MTs, endftk_file_obj = tp.extract_endf_specs(TAPE20)
         MTs = set(MTs).intersection(mt_dict.keys())
@@ -35,11 +35,10 @@ def main():
         # difference from the original MT values in the ENDF/PENDF files
         material_id, MTs, endftk_file_obj = tp.extract_endf_specs(gendf_path)
         gendf_data = tp.iterate_MTs(MTs, endftk_file_obj, mt_dict, pKZA)
-        cumulative_data = concat([cumulative_data, gendf_data],
-                                 ignore_index=True)
+        cumulative_data.extend(gendf_data)
         groupr_tools.cleanup_njoy_files()
 
-    cumulative_data.to_csv('cumulative_gendf_data.csv')
+    DataFrame(cumulative_data).to_csv('cumulative_gendf_data.csv')
 
 if __name__ == '__main__':
     main()
